portfolio-api/app/schemas/watchlist.py
from pydantic import BaseModel

class WatchlistItemBase(BaseModel):
    symbol: str

class WatchlistItemCreate(WatchlistItemBase):
    pass

class WatchlistItem(WatchlistItemBase):
    id: int
    user_id: int

    class Config:
        orm_mode = True